requirements.txt
django-ckeditor==6.5.1
django-select2==8.1.1
django-widget-tweaks==1.4.12
pillow==9.5.0
pip-chill==1.0.3
six==1.16.0
